# HOTFIX invalidate cache when modifying known predicates