--- a/booster/library/Booster/Pattern/Rewrite.hs
+++ b/booster/library/Booster/Pattern/Rewrite.hs
@@ -513,9 +513,16 @@ applyRule pat@Pattern{ceilConditions} rule =
                                                 , rulePredicate = Just rulePredicate
                                                 }
   where
-    filterOutKnownConstraints :: Set.Set Predicate -> [Predicate] -> RewriteT io [Predicate]
-    filterOutKnownConstraints priorKnowledge constraitns = do
-        let (knownTrue, toCheck) = partition (`Set.member` priorKnowledge) constraitns
+    -- These predicates are known (and do not change) during the
+    -- entire rewrite step. The simplifier cache cannot be retained
+    -- when additional predicates are used (see 'checkConstraint').
+    knownPatternPredicates =
+        pat.constraints <> (Set.fromList . asEquations $ pat.substitution)
+
+    filterOutKnownConstraints :: [Predicate] -> RewriteT io [Predicate]
+    filterOutKnownConstraints constraints = do
+        let (knownTrue, toCheck) =
+                partition (`Set.member` knownPatternPredicates) constraints
         unless (null knownTrue) $
             getPrettyModifiers >>= \case
                 ModifiersRep (_ :: FromModifiersT mods => Proxy mods) ->
@@ -534,10 +541,11 @@ applyRule pat@Pattern{ceilConditions} rule =
         Set.Set Predicate ->
         Predicate ->
         RewriteRuleAppT (RewriteT io) (Maybe a)
-    checkConstraint onUnclear onBottom knownPredicates p = do
+    checkConstraint onUnclear onBottom extraPredicates p = do
         RewriteConfig{definition, llvmApi, smtSolver} <- lift $ RewriteT ask
         RewriteState{cache} <- lift . RewriteT . lift $ get
-        simplified <-
+        let knownPredicates = knownPatternPredicates <> extraPredicates
+        simplified <- -- FIXME retain cache unless extraPredicates non-empty
             withContext CtxConstraint $
                 simplifyConstraint definition llvmApi smtSolver cache knownPredicates p
         case simplified of
@@ -554,14 +562,9 @@ applyRule pat@Pattern{ceilConditions} rule =
         -- apply substitution to rule requires
         let ruleRequires =
                 concatMap (splitBoolPredicates . substituteInPredicate matchingSubst) rule.requires
-            knownConstraints = pat.constraints <> (Set.fromList . asEquations $ pat.substitution)
 
         -- filter out any predicates known to be _syntactically_ present in the known prior
-        toCheck <-
-            lift $
-                filterOutKnownConstraints
-                    knownConstraints
-                    ruleRequires
+        toCheck <- lift $ filterOutKnownConstraints ruleRequires
 
         -- simplify the constraints (one by one in isolation). Stop if false, abort rewrite if indeterminate.
         unclearRequires <-
@@ -570,17 +573,13 @@ applyRule pat@Pattern{ceilConditions} rule =
                     ( checkConstraint
                         id
                         returnNotApplied
-                        knownConstraints
+                        mempty -- checkConstraint already considers knownConstraints
                     )
                     toCheck
 
         -- unclear conditions may have been simplified and
         -- could now be syntactically present in the path constraints, filter again
-        stillUnclear <-
-            lift $
-                filterOutKnownConstraints
-                    knownConstraints
-                    unclearRequires
+        stillUnclear <- lift $ filterOutKnownConstraints unclearRequires
 
         -- check unclear requires-clauses in the context of known constraints (priorKnowledge)
         solver <- lift $ RewriteT $ (.smtSolver) <$> ask
@@ -609,17 +608,14 @@ applyRule pat@Pattern{ceilConditions} rule =
         -- apply substitution to rule ensures
         let ruleEnsures =
                 concatMap (splitBoolPredicates . coerce . substituteInTerm matchingSubst . coerce) rule.ensures
-            knownConstraints =
-                pat.constraints
-                    <> (Set.fromList . asEquations $ pat.substitution)
-                    <> Set.fromList unclearRequiresAfterSmt
         newConstraints <-
             catMaybes
                 <$> mapM
                     ( checkConstraint
                         id
                         returnTrivial
-                        knownConstraints
+                        -- supply required path conditions as extra constraints
+                        (Set.fromList unclearRequiresAfterSmt)
                     )
                     ruleEnsures
 
@@ -667,7 +663,7 @@ applyRule pat@Pattern{ceilConditions} rule =
         let ruleRequires =
                 concatMap (splitBoolPredicates . coerce . substituteInTerm matchingSubst . coerce) rule.requires
         collapseAndBools . catMaybes
-            <$> mapM (checkConstraint id returnNotApplied pat.constraints) ruleRequires
+            <$> mapM (checkConstraint id returnNotApplied mempty) ruleRequires
 
 ruleGroupPriority :: [RewriteRule a] -> Maybe Priority
 ruleGroupPriority = \case