--- a/booster/test/rpc-integration/resources/retain-condition-cache.kompile
+++ b/booster/test/rpc-integration/resources/retain-condition-cache.kompile
@@ -1,0 +1,1 @@
+kompile-from-double-definition.sh