--- a/booster/test/rpc-integration/test-retain-condition-cache/params-before-jumpi.json
+++ b/booster/test/rpc-integration/test-retain-condition-cache/params-before-jumpi.json
@@ -1,0 +1,4 @@
+{
+    "log-successful-rewrites": true,
+    "max-depth": 1
+}