--- a/scripts/booster-integration-tests.sh
+++ b/scripts/booster-integration-tests.sh
@@ -14,7 +14,7 @@ cabal update
 cabal test llvm-integration predicates-integration
 
 # The runDirectoryTest.sh script expects the following env vars to be set
-export PLUGIN_DIR=$(nix build github:runtimeverification/blockchain-k-plugin/$PLUGIN_VERSION --no-link --json | jq -r '.[].outputs | to_entries[].value')
+export PLUGIN_DIR=$(nix build github:runtimeverification/blockchain-k-plugin/$PLUGIN_VERSION --no-write-lock-file --no-link --json | jq -r '.[].outputs | to_entries[].value')
 export SECP256K1_DIR=$(nix build nixpkgs#secp256k1 --no-link --json | jq -r '.[].outputs | to_entries[].value')
 
 cabal build all