… booster/test/rpc-integration/resources/3934-smt.kompile
… booster/test/rpc-integration/resources/foundry-bug-report.kompile
… booster/test/rpc-integration/resources/kompile-from-double-definition.sh
… booster/test/rpc-integration/test-log-simplify-json/simplify-log.txt.golden
… booster/test/rpc-integration/test-retain-condition-cache/response-before-jumpi.json
… booster/test/rpc-integration/test-retain-condition-cache/state-before-jumpi.execute
… booster/test/rpc-integration/test-substitutions/response-symbolic-bottom-predicate.json
… booster/test/rpc-integration/test-vacuous/response-vacuous-but-rewritten.json
… booster/test/rpc-integration/test-vacuous/response-vacuous-without-rewrite.json
… booster/unit-tests/Test/Booster/Pattern/ApplyEquations.hs
… deps/blockchain-k-plugin_release