The change in #3953 introduced additional arguments to evaluation functions which allow callers to supply some "known true" predicates for the simplification and evaluation. However, doing so means that the cache will get populated with associations that might only be true if this known truth does not change.

The case in point was a predicate being cached as "true" because of an earlier evaluation, and then _removed_ from the path condition (see added integration test).

This change 
* refactors `checkConstraint` functions to fix the pattern constraints
* skips the cache update if  any additional "known truth" arguments were supplied to `checkConstraint`, and
* runs internal computations that manipulate the predicates with a fresh empty cache.

Furthermore, to avoid a significant performance penalty, we
* simplify the pattern constraints once before starting to rewrite
* only simplify the pattern _term_ during rewriting
   - the pattern constraints are already evaluated and stay the same
   - if a new constraint is added (by way of an `ensures`), this constraint is evaluated before adding it
* At the end of the rewrite, we still simplify everything again (including the entire path condition).

We are not using the SMT solver to detect a vacuous state at the start, though, could add that separately if desired. At the moment, the behaviour is largely the same as before (integration test expectations identical except for some evaluation in vacuous states).