@jberthold 2025-04-14
* `kevm` tests show regression (especially when tests run longer, as expected)

| Test                                                                        | HOTFIX-invalidate-cache-when-modifying-known-predicates time | master-2a5be7830 time | (HOTFIX-invalidate-cache-when-modifying-known-predicates/master-2a5be7830) time 
|-----------------------------------------------------------------------------|--------------------------------------------------------------|-----------------------|---------------------------------------------------------------------------------
| erc20/ds/transferFrom-failure-1-c-spec.k                                    | 83.53                                                        | 80.64                 | 1.0358382936507937                                                              
| erc20/ds/transfer-failure-1-a-spec.k                                        | 57.93                                                        | 55.92                 | 1.0359442060085837                                                              
| erc20/ds/transferFrom-failure-1-a-spec.k                                    | 71.18                                                        | 68.69                 | 1.036249818023002                                                               
| erc20/ds/transfer-success-2-spec.k                                          | 47.57                                                        | 45.9                  | 1.0363834422657952                                                              
| examples/sum-to-n-foundry-spec.k                                            | 81.36                                                        | 78.42                 | 1.0374904361132364                                                              
| examples/erc20-spec.md                                                      | 114.61                                                       | 110.34                | 1.0386985680623526                                                              
| mcd-structured/flipper-bids-pass-rough-spec.k                               | 52.26                                                        | 50.31                 | 1.0387596899224805                                                              
| mcd/dsvalue-read-pass-summarize-spec.k                                      | 48.81                                                        | 46.97                 | 1.0391739408132852                                                              
| kontrol/test-expectreverttest-testfail_expectrevert_bytes4-0-spec.k         | 44.9                                                         | 43.18                 | 1.0398332561371004                                                              
| kontrol/test-expectreverttest-testfail_expectrevert_failandsuccess-0-spec.k | 44.6                                                         | 42.82                 | 1.0415693601120972                                                              
| kontrol/test-storetest-testaccesses-0-spec.k                                | 44.08                                                        | 42.23                 | 1.0438077196305944                                                              
| kontrol/test-storetest-teststoreload-0-spec.k                               | 44.1                                                         | 42.23                 | 1.044281316599574                                                               
| erc20/ds/approve-success-spec.k                                             | 44.89                                                        | 42.94                 | 1.045412203074057                                                               
| mcd-structured/flopper-tick-pass-rough-spec.k                               | 71.13                                                        | 67.84                 | 1.0484964622641508                                                              
| erc20/hkg/transferFrom-failure-1-spec.k                                     | 65.6                                                         | 62.53                 | 1.0490964337118183                                                              
| mcd-structured/flopper-kick-pass-rough-spec.k                               | 68.62                                                        | 65.36                 | 1.0498776009791921                                                              
| benchmarks/structarg01-spec.k                                               | 49.81                                                        | 47.4                  | 1.0508438818565402                                                              
| kontrol/test-safetest-testwithdrawfuzz-uint96-0-spec.k                      | 48.05                                                        | 45.7                  | 1.0514223194748358                                                              
| erc20/ds/transferFrom-success-2-spec.k                                      | 53.58                                                        | 50.93                 | 1.0520322010602787                                                              
| benchmarks/storagevar02-nooverflow-spec.k                                   | 42.47                                                        | 40.35                 | 1.052540272614622                                                               
| erc20/ds/transferFrom-success-1-spec.k                                      | 58.15                                                        | 55.01                 | 1.0570805308125795                                                              
| erc20/ds/transfer-failure-2-a-spec.k                                        | 56.47                                                        | 53.37                 | 1.0580850665167698                                                              
| erc20/ds/transferFrom-failure-2-b-spec.k                                    | 69.24                                                        | 65.43                 | 1.0582301696469507                                                              
| benchmarks/ecrecoverloop00-sig0-invalid-spec.k                              | 51.02                                                        | 48.11                 | 1.0604863853668676                                                              
| benchmarks/ecrecoverloop02-sig1-invalid-spec.k                              | 74.43                                                        | 69.85                 | 1.0655690765926988                                                              
| examples/storage-spec.md                                                    | 73.92                                                        | 68.96                 | 1.0719257540603249                                                              
| mcd/flopper-file-pass-rough-spec.k                                          | 20.03                                                        | 18.65                 | 1.0739946380697052                                                              
| erc20/ds/transfer-success-1-spec.k                                          | 50.91                                                        | 47.36                 | 1.0749577702702702                                                              
| benchmarks/ecrecoverloop02-sig0-invalid-spec.k                              | 65.51                                                        | 60.78                 | 1.0778216518591643                                                              
| benchmarks/ecrecoverloop00-sig1-invalid-spec.k                              | 78.53                                                        | 72.63                 | 1.0812336500068842                                                              
| benchmarks/ecrecoverloop00-sigs-valid-spec.k                                | 58.27                                                        | 53.86                 | 1.0818789454140365                                                              
| benchmarks/ecrecoverloop02-sigs-valid-spec.k                                | 77.47                                                        | 71.53                 | 1.0830420802460505                                                              
| mcd/flipper-ttl-pass-spec.k                                                 | 50.64                                                        | 46.55                 | 1.0878625134264233                                                              
| mcd-structured/end-pack-pass-rough-spec.k                                   | 99.41                                                        | 91.28                 | 1.0890666082383873                                                              
| mcd-structured/flapper-yank-pass-rough-spec.k                               | 82.97                                                        | 76.16                 | 1.0894170168067228                                                              
| erc20/ds/transfer-failure-1-b-spec.k                                        | 70.17                                                        | 64.33                 | 1.090781905798228                                                               
| mcd-structured/flopper-dent-guy-same-pass-rough-spec.k                      | 133.41                                                       | 121.17                | 1.101015102748205                                                               
| erc20/ds/transferFrom-failure-1-b-spec.k                                    | 108.11                                                       | 97.07                 | 1.1137323580920986                                                              
| examples/erc721-spec.md                                                     | 130.57                                                       | 116.88                | 1.1171286789869952                                                              
| mcd/flipper-tau-pass-spec.k                                                 | 52.24                                                        | 46.22                 | 1.1302466464733882                                                              
| mcd/cat-file-addr-pass-rough-spec.k                                         | 55.26                                                        | 48.78                 | 1.132841328413284                                                               
| benchmarks/ecrecover00-sigvalid-spec.k                                      | 52.78                                                        | 46.51                 | 1.134809718340142                                                               
| mcd/dsvalue-peek-pass-rough-spec.k                                          | 56.11                                                        | 49.27                 | 1.138826872336107                                                               
| mcd/dsvalue-read-pass-spec.k                                                | 56.28                                                        | 48.97                 | 1.1492750663671636                                                              
| mcd/vow-flog-fail-rough-spec.k                                              | 104.66                                                       | 90.65                 | 1.154550468836183                                                               
| mcd/dstoken-approve-fail-rough-spec.k                                       | 83.07                                                        | 71.92                 | 1.1550333704115683                                                              
| mcd-structured/pot-join-pass-rough-spec.k                                   | 143.81                                                       | 123.22                | 1.1670994968349295                                                              
| mcd-structured/flopper-dent-guy-diff-tic-not-0-pass-rough-spec.k            | 169.82                                                       | 145.49                | 1.1672279881778815                                                              
| mcd/flopper-cage-pass-spec.k                                                | 55.46                                                        | 47.43                 | 1.169302129453932                                                               
| mcd-structured/vow-fess-fail-rough-spec.k                                   | 102.08                                                       | 85.54                 | 1.1933598316577039                                                              
| mcd/vow-fess-fail-rough-spec.k                                              | 106.13                                                       | 87.48                 | 1.2131915866483767                                                              
| mcd-structured/end-cash-pass-rough-spec.k                                   | 127.27                                                       | 103.64                | 1.2280007719027402                                                              
| mcd/flopper-dent-guy-same-pass-rough-spec.k                                 | 669.44                                                       | 458.72                | 1.459365190094175                                                               
| mcd/flipper-bids-pass-rough-spec.k                                          | 100.42                                                       | 67.01                 | 1.4985823011490822                                                              
| mcd/flopper-tick-pass-rough-spec.k                                          | 214.96                                                       | 125.91                | 1.7072512111825908                                                              
| mcd/flopper-dent-guy-diff-tic-not-0-pass-rough-spec.k                       | 1077.44                                                      | 601.82                | 1.7903027483300653                                                              
| mcd/flopper-kick-pass-rough-spec.k                                          | 206.16                                                       | 107.68                | 1.914561664190193                                                               
| mcd/flapper-yank-pass-rough-spec.k                                          | 268.15                                                       | 133.35                | 2.0108736407949004                                                              
| mcd/end-pack-pass-rough-spec.k                                              | 226.9                                                        | 110.71                | 2.0494986902718817                                                              
| mcd/pot-join-pass-rough-spec.k                                              | 331.42                                                       | 158.61                | 2.0895277725238004                                                              
| mcd/end-cash-pass-rough-spec.k                                              | 368.66                                                       | 149.85                | 2.460193526860194                                                               
| TOTAL                                                                       | 7216.83                                                      | 5338.46               | 1.3518561532726667       

*  `kontrol` tests without alarming regressions

| Test                                                                                     | HOTFIX-invalidate-cache-when-modifying-known-predicates time | master-2a5be7830 time | (HOTFIX-invalidate-cache-when-modifying-known-predicates/master-2a5be7830) time 
|------------------------------------------------------------------------------------------|--------------------------------------------------------------|-----------------------|---------------------------------------------------------------------------------
| kontrol/src/tests/integration/test_foundry_prove.py::test_constructor_with_symbolic_args | 24.54                                                        | 34.22                 | 0.717124488603156                                                               
| ImmutableVarsTest.test_run_deployment(uint256)                                           | 62.0                                                         | 84.86                 | 0.7306151308036767                                                              
| StoreTest.testStoreLoadNonExistent()                                                     | 7.74                                                         | 9.56                  | 0.8096234309623431                                                              
| BytesTypeTest.test_bytes4(bytes4)                                                        | 7.38                                                         | 8.69                  | 0.8492520138089759                                                              
| ExpectCallTest.testExpectStaticCall()                                                    | 10.67                                                        | 12.25                 | 0.8710204081632653                                                              
| StoreTest.testGasLoadWarmUp()                                                            | 10.19                                                        | 11.59                 | 0.8792062122519413                                                              
| ExpectRevertTest.testFail_expectRevert_false()                                           | 11.3                                                         | 12.57                 | 0.8989657915672236                                                              
| EmitContractTest.testExpectEmit()                                                        | 9.96                                                         | 11.03                 | 0.9029918404351769                                                              
| ExpectRevertTest.test_expectRevert_returnValue()                                         | 14.39                                                        | 15.29                 | 0.9411379986919556                                                              
| SetUpDeployTest.test_extcodesize()                                                       | 22.31                                                        | 23.41                 | 0.9530115335326783                                                              
| PlainPrankTest.test_startPrankWithOrigin_true()                                          | 11.61                                                        | 12.02                 | 0.9658901830282862                                                              
| MockFunctionTest.test_mock_function_concrete_args()                                      | 24.17                                                        | 23.35                 | 1.035117773019272                                                               
| Setup2Test.testFail_setup()                                                              | 16.06                                                        | 15.51                 | 1.0354609929078014                                                              
…